% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-calculate.R
\name{calculate_exposure}
\alias{calculate_exposure}
\title{Calculate exposure to active ingredients present in applied pesticides.}
\usage{
calculate_exposure(clean_pur_df, location, radius, time_period = NULL,
  start_date = NULL, end_date = NULL, chemicals = "all")
}
\arguments{
\item{clean_pur_df}{A data frame returned by \code{clean_pur_data} that
includes data for the county in which your location is located (before
running \code{clean_pur_data}, you can use the \code{find_location_county}
function to figure this out), the time period, and the active ingredients
or chemical classes for which you want to calculate exposure.}

\item{location}{A character string. Either a California address including
street name, city, state, and zip code, or a pair of coordinates in the
form "longitude, latitude".}

\item{radius}{A numeric value greater than zero that gives the radius in meters
defining the buffer around your location in which you would like to
calculate exposure. For reference, the length and width of a PLS section is
about 1,609 meters (1 mile). That of a township could range from about
9,656 to 11,265 meters (6-7 miles).}

\item{time_period}{Optional. A character string giving a time period over which you
would like to calculate exposure. For example, if you enter "6 months" for
\code{time_period}, \code{calculate_exposure} will calculate exposure for
every six month period starting from the earliest date present in the
\code{clean_pur_df} data frame. Start and end dates can be optionally specified
with the \code{start_date} and \code{end_date} arguments. Alternatively, to
calculate exposure over only one time period, you can leave this argument
NULL and specify start and end dates.}

\item{start_date}{Optional. "yyyy-mm-dd" specifying the start date for
exposure estimation. This date should be present in the \code{clean_pur_df}
data frame.}

\item{end_date}{Optional. "yyyy-mm-dd" specifying the end date for exposure
estimation. This date should be present in the \code{clean_pur_df}
data frame.}

\item{chemicals}{A character string indicating active ingredients for which you
would like to calculate exposure. The default is "all", which will calculate
exposure to the summed active ingredients present in the \code{clean_pur_df}
data frame. You can also enter a particular active ingredient that's present
in your \code{clean_pur_df} data frame, or "chemical_class" to calculate
exposure to each of the chemical classes present in the \code{chemical_class}
column of your \code{clean_pur_df} data frame.}
}
\value{
A list with three elements:
 \describe{
   \item{exposure}{A data frame with a row for each date
   range for which exposure was calculated and seven columns: \code{exposure},
   the estimate of exposure in kg/m^2, \code{location}, \code{start_date} and
   \code{end_date}, \code{radius}, the radius in meters for the buffer
   extending from the location, and \code{chemcials}, (either "all", indicating
   that all active ingredients present in the \code{clean_pur_df} were summed,
   a particular active ingredient, or the chemical class(es) specified in the
   \code{clean_pur_df} data frame). \code{none_recorded} is TRUE if exposure was
   0 because there was no pesticide application recorded in the sections or
   townships intersected by the buffer for that date range.}
   \item{meta_data}{A data frame with 11 columns and one row for every section
   or township intersected by the specified buffer extending from the
   given location. Columns include \code{location}, \code{section} or
   \code{township}, \code{percent}, the percent that the section is overlapped
   by the buffer, \code{kg}, the total amount of kg applied for the specified
   chemicals and date range in that section or township, \code{kg_int}, the
   amount of kilograms applied multiplied by the percent of overlap,
   \code{start_date} and \code{end_date}, \code{radius}, \code{chemicals},
   \code{area}, the area of the buffer (m^2), and \code{none_recorded}, logical
   for whether any pesticide application was recorded for the specified section
   or township, date range, and chemicals.}
   \item{buffer_plot}{A data frame with 24 columns. Contains spatial plotting
   data for the buffer and overlapping sections or townships. You can use the
   \code{plot_tidy} function to quickly plot and get a rough idea of the
   area for which exposure was calculated, before moving on to other map_*
   or plot_* functions.}
 }
}
\description{
For a particular location, buffer radius, date range, and active ingredient
or class of active ingredients, \code{calculate_exposure} calculates an
estimate of exposure in kg of active ingredient per m^2.
}
\section{Note}{

 \itemize{
   \item{If the \code{time_period}, \code{start_date}, and \code{end_date}
         arguments are all left as NULL (their defaults), then exposure will
         be estiamted across the entire date range of the \code{clean_pur_df}
         data frame.}
   \item{If you pulled PUR data from \code{clean_pur_data} specifying
         \code{sum_application = TRUE} and \code{unit = "township"}, then
         exposure will be calculated based on townships. Using the
         \code{plot_tidy} function to plot the returned \code{buffer_plot}
         list element could be helpful to see the difference between
         calculating exposure based on sections or townships for a certain
         buffer radius.}
}
}

\examples{
\dontrun{
clean_pur <- clean_pur_data(2000, "10", "all", TRUE)

pur_data <- clean_pur_data()

# Time periods

# Time periods with specified start and end date

# Single time period

# All active ingredients

# A single active ingredient

# By chemical class
pur_data_classes <- clean_pur_data()

}
}
