% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{california_shp}
\alias{california_shp}
\title{California state SpatialPolygonsDataFrame object.}
\format{A SpatialPolygonsDataFrame object at the 1:20,000,000 resolution level.}
\source{
\url{https://www.census.gov/geo/maps-data/data/cbf/cbf_state.html}
}
\usage{
california_shp
}
\description{
A SpatialPolygonsDataFrame object for the outline of the state of California.
Downloaded and subset from the 2016 U.S. Census Cartographic Boundary
Shapefile for states.
}
\keyword{datasets}
