% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_pur.R
\name{chemical_codes}
\alias{chemical_codes}
\title{Pull active ingredient chemical codes from PUR Chemical Lookup Tables.}
\usage{
chemical_codes(year, chemicals = "all")
}
\arguments{
\item{year}{A four-digit numeric year in the range of 1990 to 2015. Indicates
the year in which you would like to match chemical codes.}

\item{chemicals}{A string or vector of strings giving search terms of
chemicals to match with active ingredients present in pesticides applied
in the given year. The default value is "all", which returns codes for all
active ingredients applied in a given year.}
}
\value{
A data frame with three columns:
  \itemize{
    \item \code{chem_code}, with chemical codes corresponding to each active
    ingredient. \code{chem_code} values are used to later filter raw PUR
    datasets.
    \item \code{chemname}, with unique active ingredients corresponding to
    each search term.
    \item \code{chemical}, with search terms given in the \code{chemicals}
    argument.
    }
}
\description{
For a vector of chemical names, \code{chemical_codes} returns
a data frame with corresponding chemical codes from the PUR Chemical Lookup
Table for a given year.
}
\section{Note}{

The PUR Chemical Lookup Table for a year lists all active ingredients present
in applied pesticides across the state of California. Therefore, PUR data for
a particular county may not include records for active ingredients returned
by \code{chemical_search} for the same year.
}

\examples{
\dontrun{
chemical_codes(2000, "methyl bromide")
chemical_codes(1995, c("ammonia", "benzene"))
}
}
