% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chemical_list}
\alias{chemical_list}
\title{California Pesticide Use Report chemical codes.}
\format{A list of 26 elements. Each element is a data frame with a variable
number of rows ranging from 3,579 to 3,934 and two columns:
\describe{
  \item{chem_code}{An integer giving the chemical code. This uniquely
  identifies a chemical within a year.}
  \item{chemname}{A character vector giving common chemical name for each
  active ingredient. These are usually listed on the pesticide product
  label.}
}}
\source{
\url{ftp://transfer.cdpr.ca.gov/pub/outgoing/pur_archives}
}
\usage{
chemical_list
}
\description{
A list of data frames (one for each year from 1990 through 2017) containing
California Department of Pesticide Regulation chemical codes and names used
to identify active ingredients in Pesticide Use Report data. "chemical.txt"
files for each year were pulled from the .zip files "pur1990.zip" through
"pur2017.zip" found here:
\url{ftp://transfer.cdpr.ca.gov/pub/outgoing/pur_archives}
}
\keyword{datasets}
