% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_pur.R
\name{chemical_search}
\alias{chemical_search}
\title{Search for active ingredients applied in a year.}
\usage{
chemical_search(year, chemicals)
}
\arguments{
\item{year}{A four-digit numeric year in the range of 1990 to 2015. Indicates
the year in which you would like to search for active ingredients.}

\item{chemicals}{A string or vector of strings giving search terms of
chemicals to match with active ingredients present in pesticides applied
in the given year.}
}
\value{
A list of data frames (one data frame per search term). Each data
  frame has two columns: \code{chemname}, giving the active ingredient
  present in PUR records for that year, and \code{search_term}, giving the
  corresponding search term provided in the \code{chemicals} argument.
  Elements of the list are named according to given search terms.
}
\description{
\code{chemical_search} returns a list of data frames with active ingredients
(\code{chemname}) matching each search term given in the \code{chemicals}
argument for a particular year.
}
\section{Note}{

The PUR Chemical Lookup Table for a year lists all active ingredients present
in applied pesticides across the state of California. Therefore, PUR data for
a particular county may not include records for active ingredients returned
by \code{chemical_search} for the same year.
}

\examples{
\dontrun{
chemical_search(2000, "methyl bromide")
chemical_search(1995, c("ammonia", "benzene"))
}
}
