% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{county_codes}
\alias{county_codes}
\title{California Pesticide Use Report county codes.}
\format{A data frame with 58 rows and two columns:
\describe{
  \item{county_name}{A character vector giving California county names.}
  \item{county_code}{A character vector giving county codes (ranging from "01"
  through "58") corresponding to each California county. Note: these codes
  are unique to California PUR datasets; they do not correspond to FIPS
  codes.}
}}
\source{
\url{ftp://transfer.cdpr.ca.gov/pub/outgoing/pur_archives}
}
\usage{
county_codes
}
\description{
A data frame containing California county names and corresponding codes
used to identify counties in California Pesticide Use Reports. This file,
"county.txt", was pulled from the .zip file "pur2000.zip" found here:
\url{ftp://transfer.cdpr.ca.gov/pub/outgoing/pur_archives}
}
\keyword{datasets}
