% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-helpers.R
\name{exp_out_val}
\alias{exp_out_val}
\title{Return a single exposure value for each combination of conditions.}
\usage{
exp_out_val(...)
}
\arguments{
\item{...}{Either \code{chemicals} or \code{chemicals, aerial_ground}. Not
quoted.}
}
\value{
A data frame with exposure values in kg/m^2 at a location for each
relevant condition.
}
\description{
\code{exp_out_val} returns a data frame with exposure values (kg/m^2) by
chemicals (including \code{chemicals = "all"}) or by chemicals and aerial/ground
application.
}
\details{
This is a helper function for \code{daterange_calcexp}.
}
