% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exposure_ex2}
\alias{exposure_ex2}
\title{Another toy exposure list}
\format{A list with four data frame elements: `exposure`, `meta_data`,
`buffer_plot_df`, and `clean_pur_df`.}
\usage{
exposure_ex2
}
\description{
Another list created using the `calculate_exposure` function for the
`fresno_ex` dataset, the location (longitude, latitude)
"-119.247100, 37.204875", and a radius of 3,000 meters. The `county_plot`
list element has been removed.
}
\keyword{datasets}
