% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-find.R
\name{find_counties}
\alias{find_counties}
\title{Find California county codes or names.}
\usage{
find_counties(counties, return = "pur_codes")
}
\arguments{
\item{counties}{A vector of character strings giving either a county names,
two digit PUR county codes, or six-digit FIPS county codes. Not case
sensitive. California names and county codes as they appear in PUR data sets
can be found in the \code{county_codes} data set available with this package.}

\item{return}{Either "pur_codes" to return PUR county codes (the default),
"fips_codes" to return FIPS county codes, or "names" to return county names.}
}
\value{
If \code{return = "pur_codes"}, a vector of two-character strings giving
  the corresponding PUR county codes. If \code{return = "fips_codes"}, a vector
  of six-digit character strings giving the corresponding FIPS county codes.
  If \code{return = "names"}, a vector of county names.
}
\description{
Given a vector of counties, \code{find_counties} returns either PUR
county codes or names.
}
\examples{
find_counties(c("01", "06005", "el dorado"))
\donttest{
find_counties(c("01", "06005", "el dorado"), return = "fips_codes")
find_counties(c("01", "06005", "el dorado"), return = "names")
}
}
