% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_pur.R
\name{find_county_code}
\alias{find_county_code}
\title{Find California county codes}
\usage{
find_county_code(county)
}
\arguments{
\item{county}{A character string giving either a county name or a two digit
PUR county code. Not case sensitive. California names and county codes as
they appear in PUR datasets can be found in the county_codes dataset
available with this package. For example, to return data for Alameda county,
enter either "alameda" or "01" for the county argument.}
}
\value{
A two-character string giving the corresponding PUR county code.
}
\description{
Given a county, \code{find_county_code} returns a PUR county code.
}
\examples{
\dontrun{
find_county_code("el dorado")
find_county_code("45")
}
}
