% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{find_county_codes}
\alias{find_county_codes}
\title{Find California county codes}
\usage{
find_county_codes(counties)
}
\arguments{
\item{county}{A vector of character strings giving either a county names or
two digit PUR county codes. Not case sensitive. California names and county
codes as they appear in PUR datasets can be found in the county_codes dataset
available with this package.}
}
\value{
A vector of two-character string giving the corresponding PUR county
  codes.
}
\description{
Given a vector of counties, \code{find_county_codes} returns PUR county codes.
}
\examples{
find_county_codes(c("01", "03", "el dorado"))
find_county_codes(c("contra costa", "45"))
}
