% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-helpers.R
\name{find_cutpoints}
\alias{find_cutpoints}
\title{Categorize a continuous scale by percentile cutpoints}
\usage{
find_cutpoints(section_data, buffer_or_county, start_date = NULL,
  end_date = NULL, aerial_ground = NULL, chemicals = NULL,
  clean_pur = NULL, s_t = NULL, percentile)
}
\arguments{
\item{section_data}{A data frame with a continuous numeric variable ("kg")}

\item{buffer_or_county}{Should cutpoints be determined by "county" or "buffer"?}

\item{start_date}{A date}

\item{end_date}{A date}

\item{clean_pur}{A \code{pull_clean_pur} data frame}

\item{s_t}{"section" or "township"}

\item{percentile}{A numeric vector in (0, 1).}
}
\value{
The input \code{section_data} data frame with an additional column
named \code{category}.
}
\description{
Given a data frame with a column with a continuous numeric variable,
\code{find_cutpoints} returns the data frame with a new column categorizing
that continuous variable by percentile cutpoints.
}
\details{
This is a helper function for \code{plot_pls}.
}
