% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-find.R
\name{find_product_name}
\alias{find_product_name}
\title{Find Pesticide Product names and registration numbers from PUR Product Lookup
Tables.}
\usage{
find_product_name(year, products = "all", download_progress = FALSE)
}
\arguments{
\item{year}{A four digit year in the range of 1990 to 2015.}

\item{products}{A character string or a vector of character strings with
pesticide product names that you would like to search for. Not case
sensitive. The default is "all", which will return all pesticide products
applied for a given year.}

\item{download_progress}{TRUE / FALSE indicating whether you would like a
message and progress bar printed for the product table that is downloaded.
The default value is FALSE.}
}
\value{
A data frame with six columns:
\describe{
  \item{prodno}{The CA registration number. Can be matched with the
  \code{prodno} in a raw or cleaned PUR dataset.}
  \item{prodstat_ind}{Character. An indication of product registration status:
  \itemize{
  \item A = Active
  \item B = Inactive
  \item C = Inactive, Not Renewed
  \item D = Inactive, Voluntary Cancellation
  \item E = Inactive, Cancellation
  \item F = Inactive, Suspended
  \item G = Inactive, Invalid Data
  \item H = Active, Suspended}}
  \item{product_name}{Character. The name of the product taken from the
  registered product label. May have been modified by DPR's Registration Branch
  to ensure uniqueness.}
  \item{signlwrd_ind}{Integer. The signal word printed on the front of the
product label:
  \itemize{
  \item 1 = Danger (Poison)
  \item 2 = Danger (Only)
  \item 3 = Warning
  \item 4 = Caution
  \item 5 = None}}
  \item{year}{The year for which product table information was pulled.}
  \item{product}{Product name search terms.}
}
}
\description{
For a year and vector of product search terms, \code{find_product_name} returns
a data frame with corresponding product registration numbers, \code{prodno},
indicator codes, and product names.
}
\examples{
\dontrun{
prod_df <- find_product_name(2000, "mosquito")
prod_df2 <- find_product_name(2010, c("insecticide", "rodenticide"))
}
}
