% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fresno_clean}
\alias{fresno_clean}
\title{An example clean PUR dataset}
\format{A data frame with 16,907 rows and 13 columns.}
\usage{
fresno_clean
}
\description{
A PUR dataset pulled using `pull_clean_pur` for Fresno county for the year
2000 and filtered to the month of January.
}
\keyword{datasets}
