% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fresno_raw}
\alias{fresno_raw}
\title{An example raw PUR dataset}
\format{A data frame with 18,062 rows and 33 columns.}
\usage{
fresno_raw
}
\description{
A PUR dataset pulled using `pull_raw_pur` for Fresno county for the year
2000 and filtered to the month of January.
}
\keyword{datasets}
