% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fresno_spdf}
\alias{fresno_spdf}
\title{SpatialPolygonsDataFrame describing Fresno county sections}
\format{An object of class \code{SpatialPolygonsDataFrame} with 6190 rows and 17 columns.}
\usage{
fresno_spdf
}
\description{
A SpatialPolygonsDataFrame object pulled using the `pull_spdf` function for
Fresno county sections.
}
\keyword{datasets}
