% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-misc.R
\name{gradient_n_pal2}
\alias{gradient_n_pal2}
\title{Include alpha option in scales::gradient_n_pal().}
\usage{
gradient_n_pal2(colours, values = NULL, space = "Lab", alpha = NULL)
}
\description{
This function adds an "alpha" argument from gradient_n_pal() from the scales
package.
}
