% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-helpers.R
\name{help_calc_exp}
\alias{help_calc_exp}
\title{Return a single exposure value for each combination of conditions.}
\usage{
help_calc_exp(exp, buffer_area, ...)
}
\arguments{
\item{exp}{A data frame}

\item{...}{Either \code{chemicals} or \code{chemicals, aerial_ground}. Not
quoted}
}
\value{
A data frame with exposure values in kg/m^2 at a location for each
relevant condition.
}
\description{
\code{help_calc_exp} returns a data frame with exposure values (kg/m^2) by
chemicals (including \code{chemicals = "all"}) or by chemicals and aerial/ground
application.
}
\details{
This is a helper function for \code{help_calculate_exposure}.
}
