% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-helpers.R
\name{help_calculate_exposure}
\alias{help_calculate_exposure}
\title{Return exposure data for a single start and end date}
\usage{
help_calculate_exposure(start_date, end_date, aerial_ground, chemicals,
  clean_pur_df, location, pls_percents, pur_filt, radius)
}
\arguments{
\item{start_date}{Optional. "yyyy-mm-dd" specifying the start date for
exposure estimation. This date should be present in the \code{clean_pur_df}
data frame.}

\item{end_date}{Optional. "yyyy-mm-dd" specifying the end date for exposure
estimation. This date should be present in the \code{clean_pur_df}
data frame.}

\item{aerial_ground}{TRUE / FALSE for whether you would like to
incorporate aerial/ground application into exposure calculations. If
\code{aerial_ground = TRUE}, there should be an \code{aerial_ground}
column in the input \code{clean_pur_df} data frame. There will be a value of
exposure calculated for each chemical ("all" or by chemical class) and for
each method of application: aerial or ground. The default is FALSE.}

\item{chemicals}{Either "all" or "chemical_class". The default is "all", which
will calculate exposure to the summed active ingredients present in the
\code{clean_pur_df} data frame. Enter "chemical_class" to calculate
exposure to each of the chemical classes present in the \code{chemical_class}
column of your \code{clean_pur_df} data frame.}

\item{clean_pur_df}{A data frame returned by \code{pull_clean_pur} that
includes data for the county of your location (before running
\code{pull_clean_pur}, you can use the \code{find_location_county}
function to figure this out), the time period, and the active ingredients
or chemical classes for which you want to calculate exposure.}

\item{location}{A length-one character string. Either a California address
including street name, city, state, and zip code, or a pair of coordinates
in the form "longitude, latitude".}

\item{pls_percents}{A data frame}

\item{pur_filt}{A data frame}

\item{radius}{A numeric value greater than zero that gives the radius in meters
defining the buffer around your location in which you would like to
calculate exposure. For reference, the length and width of a PLS section is
about 1,609 meters (1 mile). That of a township could range from about
9,656 to 11,265 meters (6-7 miles).}
}
\value{
A nested data frame with two columns: The \code{row_out} column
contains the \code{exposure} data frame for the date range, and
\code{meta_data} contains the \code{meta_data} data frame for the date range.
}
\description{
For a single date range, \code{help_calculate_exposure} returns the
\code{exposure} and \code{meta_data} data frames to be output in the
\code{calculate_exposure} list as a nested data frame.
}
\details{
This is a helper function for \code{calculate_exposure}.
}
