% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-helpers.R
\name{help_categorize}
\alias{help_categorize}
\title{Categorize a continuous scale by percentile cutpoints}
\usage{
help_categorize(section_data, buffer_or_county, start_date = NULL,
  end_date = NULL, aerial_ground = NULL, chemicals = NULL,
  clean_pur = NULL, s_t = NULL, percentile)
}
\arguments{
\item{section_data}{A data frame with a continuous numeric variable named "kg"}

\item{buffer_or_county}{Should cutpoints be determined by "county" or "buffer"?}

\item{start_date}{Optional. "yyyy-mm-dd" specifying the start date for
exposure estimation. This date should be present in the \code{clean_pur_df}
data frame.}

\item{end_date}{Optional. "yyyy-mm-dd" specifying the end date for exposure
estimation. This date should be present in the \code{clean_pur_df}
data frame.}

\item{aerial_ground}{TRUE / FALSE for whether you would like to
incorporate aerial/ground application into exposure calculations. If
\code{aerial_ground = TRUE}, there should be an \code{aerial_ground}
column in the input \code{clean_pur_df} data frame. There will be a value of
exposure calculated for each chemical ("all" or by chemical class) and for
each method of application: aerial or ground. The default is FALSE.}

\item{chemicals}{Either "all" or "chemical_class". The default is "all", which
will calculate exposure to the summed active ingredients present in the
\code{clean_pur_df} data frame. Enter "chemical_class" to calculate
exposure to each of the chemical classes present in the \code{chemical_class}
column of your \code{clean_pur_df} data frame.}

\item{clean_pur}{A \code{pull_clean_pur} data frame}

\item{s_t}{"section" or "township"}

\item{percentile}{A numeric vector in (0, 1) specifying percentile cutpoints
if \code{color_by = "percentile"}. The default is \code{c(0.25, 0.5, 0.75)},
which results in four categories: < 25th percentile, >= 25th to < 50th,
>= 50th to < 75th, and >= 75th.}
}
\value{
The input \code{section_data} data frame with an additional column
named \code{category}.
}
\description{
Given a data frame with a column with a continuous numeric variable,
\code{help_categorize} returns the data frame with a new column categorizing
that continuous variable by percentile cutpoints.
}
\details{
This is a helper function for \code{help_map_exp}.
}
