% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-helpers.R
\name{help_euc_distance}
\alias{help_euc_distance}
\title{Calculate euclidean distance between two points.}
\usage{
help_euc_distance(long, lat, origin_long, origin_lat)
}
\arguments{
\item{long}{Longitude (x) of second point}

\item{lat}{Latitude (y) of second point}

\item{origin_long}{Longitude (x) of first point}

\item{origin_lat}{Latitude (y) of first point}
}
\value{
A data frame with one row and three columns: \code{long} and
\code{lat} give the second point's coordinates, and \code{dist} gives the
euclidian distance from these coordinates from the origin.
}
\description{
\code{help_euc_distance} calculates the straight-line distance between
two points.
}
\details{
This is a helper function for \code{calculate_exposure}.
}
