% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-helpers.R
\name{help_filter_pls}
\alias{help_filter_pls}
\title{Find PLS units intersecting with a buffer.}
\usage{
help_filter_pls(pls, pls_quote, which_pls, shp, buffer, df, clean_pur_df)
}
\arguments{
\item{pls}{Either \code{MTRS} (sections) or \code{MTR} (townships). Not quoted}

\item{pls_quote}{Either \code{"MTRS"} or \code{"MTR"}}

\item{which_pls}{A vector of character string PLS units}

\item{shp}{A county's shapefile}

\item{buffer}{A data frame with buffer coordinates}

\item{df}{A data frame}

\item{clean_pur_df}{A data frame returned by \code{pull_clean_pur} that
includes data for the county of your location (before running
\code{pull_clean_pur}, you can use the
\code{find_location_county} function to figure this out), the time period,
and the active ingredients or chemical classes for which you want to
calculate exposure.}
}
\value{
A list with four elements:
\describe{
  \item{pur_filt}{A cleaned PUR data frame filtered to PLS units intersecting
  with a buffer.}
  \item{comb_df_filt}{A spatial data frame with intersecting PLS units and
  a buffer.}
  \item{pls_intersections}{A data frame with two columns: \code{pls} and
  \code{percent}, the corresponding percent intersection with the buffer.}
  \item{pls_int}{A character vector with all PLS units intersecting with the
  buffer.}
}
}
\description{
\code{help_filter_pls} filters a SpatialPolygonsDataFrame to include only PLS
units intersecting with a buffer, and filters the data frame returned from
\code{pull_clean_pur} to include only those sections or townships.
}
\details{
This is a helper function for \code{calculate_exposure}.
}
