% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-helpers.R
\name{help_find_product}
\alias{help_find_product}
\title{Pull a chemical code based on its name.}
\usage{
help_find_product(product, df)
}
\arguments{
\item{df}{A chemical table data frame.}

\item{chemical}{A string giving search term of a
chemical to match with active ingredients present in pesticides applied
in the given year.}
}
\value{
A data frame with three columns: \code{chem_code}, \code{chemname},
and \code{chemical}.
}
\description{
This function uses grep to return \code{chemname} values that match an input
search term.
}
\details{
This is a helper function for \code{find_chemical_codes}.
}
