% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-helpers.R
\name{help_map_exp}
\alias{help_map_exp}
\title{Return a map for a given exposure estimate.}
\usage{
help_map_exp(start_date, end_date, chemicals, aerial_ground, none_recorded,
  data_pls, gradient, location_longitude, location_latitude, buffer_df, buffer2,
  buffer, buffer_or_county, alpha, clean_pur, pls_labels, pls_labels_size,
  percentile, color_by)
}
\arguments{
\item{start_date}{Optional. "yyyy-mm-dd" specifying the start date for
exposure estimation. This date should be present in the \code{clean_pur_df}
data frame.}

\item{end_date}{Optional. "yyyy-mm-dd" specifying the end date for exposure
estimation. This date should be present in the \code{clean_pur_df}
data frame.}

\item{chemicals}{Either "all" or "chemical_class". The default is "all", which
will calculate exposure to the summed active ingredients present in the
\code{clean_pur_df} data frame. Enter "chemical_class" to calculate
exposure to each of the chemical classes present in the \code{chemical_class}
column of your \code{clean_pur_df} data frame.}

\item{aerial_ground}{TRUE / FALSE for whether you would like to
incorporate aerial/ground application into exposure calculations. If
\code{aerial_ground = TRUE}, there should be an \code{aerial_ground}
column in the input \code{clean_pur_df} data frame. There will be a value of
exposure calculated for each chemical ("all" or by chemical class) and for
each method of application: aerial or ground. The default is FALSE.}

\item{data_pls}{A data frame}

\item{gradient}{A character vector of hex color codes}

\item{location_longitude}{A numeric longitude value}

\item{location_latitude}{A numeric latitude value}

\item{buffer_df}{A data frame}

\item{buffer2}{A data frame}

\item{buffer}{A gpc.poly object}

\item{buffer_or_county}{"buffer" or "county"}

\item{alpha}{A number in [0,1] specifying the transperency of fill colors.
Numbers closer to 0 will result in more transparency. The default is 0.7.}

\item{clean_pur}{A \code{clean_pur_df} data frame}

\item{pls_labels}{TRUE / FALSE for whether you would like sections or townships
to be labeled with their PLS ID. The default is \code{FALSE}.}

\item{pls_labels_size}{A number specifying the size of PLS labels. The default
is 4.}

\item{percentile}{A numeric vector in (0, 1) specifying percentile cutpoints
if \code{color_by = "percentile"}. The default is \code{c(0.25, 0.5, 0.75)},
which results in four categories: < 25th percentile, >= 25th to < 50th,
>= 50th to < 75th, and >= 75th.}

\item{color_by}{Either "amount" (the default) or "percentile". Specifies
whether you would like application amounts to be colored according to
amount, resulting in a gradient legend, or by the percentile that they fall
into for the given dataset and date range. You can specify percentile
cutpoints with the \code{percentile} argument.}
}
\value{
A list with three elements:
\describe{
  \item{plot}{An exposure plot}
  \item{data}{A data frame with information about each PLS unit}
  \item{cutoff_values}{A data frame with cutoff values returned if
  \code{color_by = "percentile"}}
}
}
\description{
For a unique combination of time periods, chemicals, and application methods,
\code{help_map_exp} returns a plot showing amounts of pesticides applied for PLS
units intersecting with the buffer.
}
\details{
This is a helper function for \code{plot_exposure}.
}
