% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-helpers.R
\name{help_pull_pur}
\alias{help_pull_pur}
\title{Pull raw PUR file for a single year and a county or counties.}
\usage{
help_pull_pur(year, counties = "all", download_progress = TRUE)
}
\arguments{
\item{year}{A four-digit numeric year in the range of
1990 to 2015. Indicates the year for which you would like to pull PUR
datasets.}

\item{counties}{A character vector giving either county names or two digit
county codes. Not case sensitive. California names and county codes as they
appear in PUR datasets can be found in the county_codes dataset available
with this package. For example, to return data for Alameda county, enter
either "alameda" or "01" for the county argument. \code{counties = "all"}
will pull data for all 58 California counties.}

\item{download_progress}{TRUE / FALSE indicating whether you would like a
message and progress bar printed for each year and county of PUR data that
is downloaded. The default value is TRUE.}
}
\value{
A data frame with 33 columns. Counties are indicated by
  \code{county_cd}; the year for which data was pulled is indicated by
  \code{applic_dt}.
}
\description{
\code{help_pull_pur} pulls the raw PUR dataset for a particular year and
saves datasets for specified counties in a data frame.
}
\details{
This is a helper function for \code{pull_raw_pur}.
}
\section{Note}{

If this function returns an error (because the FTP site is down, for
example), check your working directory. You may want to change it back from a
temporary directory.
}

