% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-helpers.R
\name{help_read_in_counties}
\alias{help_read_in_counties}
\title{Read in PUR county dataset for a year.}
\usage{
help_read_in_counties(code_or_file, type, year)
}
\arguments{
\item{code_or_file}{A PUR county code or a file name for a county's dataset.}

\item{type}{Either "codes" or "files", specifying the type of argument supplied
to \code{code_or_file}.}
}
\value{
A data frame of raw PUR data for a single county and year.
}
\description{
Once a year of PUR data has been downloaded, this function reads in a
selected county's dataset.
}
\details{
This is a helper function for \code{pull_pur_file}.
}
