% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-helpers.R
\name{help_return_exposure}
\alias{help_return_exposure}
\title{Return a data frame with exposure values and other related data.}
\usage{
help_return_exposure(start_date, end_date, location, radius, exp, buffer_area,
  ...)
}
\arguments{
\item{start_date}{Optional. "yyyy-mm-dd" specifying the start date for
exposure estimation. This date should be present in the \code{clean_pur_df}
data frame.}

\item{end_date}{Optional. "yyyy-mm-dd" specifying the end date for exposure
estimation. This date should be present in the \code{clean_pur_df}
data frame.}

\item{location}{A length-one character string. Either a California address
including street name, city, state, and zip code, or a pair of coordinates
in the form "longitude, latitude".}

\item{radius}{A numeric value greater than zero that gives the radius in meters
defining the buffer around your location in which you would like to
calculate exposure. For reference, the length and width of a PLS section is
about 1,609 meters (1 mile). That of a township could range from about
9,656 to 11,265 meters (6-7 miles).}

\item{exp}{A data frame}

\item{...}{Either \code{chemicals} or \code{chemicals, aerial_ground}. Not
quoted.}
}
\value{
A data frame with one row and the columns found in the
\code{calculate_exposure$exposure} data frame.
}
\description{
For a single date range, \code{help_return_exposure} returns a data frame with
exposure values calculated from \code{help_calc_exp} as well as other relevant
data. This one row data frame is combined with data frames for other date
ranges and then returned as the \code{exposure} element from a
\code{calculate_exposure} list.
}
\details{
This is a helper function for \code{help_calculate_exposure}.
}
