% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-helpers.R
\name{help_sum_ai}
\alias{help_sum_ai}
\title{Find summed applied active ingredients}
\usage{
help_sum_ai(pur_filt, start_date, end_date, ...)
}
\arguments{
\item{pur_filt}{A data frame}

\item{start_date}{Optional. "yyyy-mm-dd" specifying the start date for
exposure estimation. This date should be present in the \code{clean_pur_df}
data frame.}

\item{end_date}{Optional. "yyyy-mm-dd" specifying the end date for exposure
estimation. This date should be present in the \code{clean_pur_df}
data frame.}

\item{...}{A list of variables to group by. Options include \code{section},
\code{township}, \code{chemical_class}, and \code{aerial_ground}. Not quoted}
}
\value{
A data frame a \code{kg} column and one to three additional columns,
depending on the grouping variables.
}
\description{
\code{help_sum_ai} finds the summed amount of applied active ingredients by
section or township, chemical class, and aerial/ground application.
}
\details{
This is a helper function for \code{help_calculate_exposure}.
}
