% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-helpers.R
\name{help_sum_application}
\alias{help_sum_application}
\title{Sum application by section, township, chemical, and method of application}
\usage{
help_sum_application(df, sum, unit, aerial_ground, section_townships,
  chemical_class = NULL, ...)
}
\arguments{
\item{df}{A data frame from \code{pull_clean_pur} before summing has taken
place}

\item{sum}{A character string giving either "all" (the
default) or "chemical_class". If \code{sum_application = TRUE},
\code{sum} indicates whether you would like to sum across all active
ingredients, giving an estimation of the total pesticides applied in a
given section or township ("all"), or by a chemical class specified in a
data frame given in the argument \code{chemical_class}.}

\item{unit}{A character string giving either "section" or "township".
Specifies whether applications of each active ingredient should be summed
by California section (the default) or by township. Only used if
\code{sum_application} is \code{TRUE}.}

\item{aerial_ground}{TRUE / FALSE indicating if you would like to
retain aerial/ground application data ("A" = aerial, "G" = ground, and
"O" = other.) The default is FALSE.}

\item{section_townships}{A section_townships data frame}

\item{chemical_class}{A data frame with only three columns: \code{chem_code},
\code{chemname}, and \code{chemical_class}. \code{chem_code} should have
integer values giving PUR chemical codes, and \code{chemname} should have
character strings with corresponding PUR chemical names (these can be
searched for using the \code{find_chemical_codes} function or with the
\code{chemical_list} dataset included with this package). The
\code{chemical_class} column should have character strings indicating the
chemical class corresponding to each \code{chem_code}. The
\code{chemical_class} for a group of active ingredients should be decided
upon by the user. Only used if \code{sum = "chemical_class"}. See
the CDPR's Summary of PUR Data document here:
\url{http://www.cdpr.ca.gov/docs/pur/pur08rep/chmrpt08.pdf} for
comprehensive classifications of active ingredients.}

\item{...}{grouping variables}
}
\value{
A data frame. The number of columns is dependent on the grouping
variables supplied to the \code{...} argument.
}
\description{
\code{help_sum_application} sums application of a PUR dataset by chemicals,
PLS unit, and aerial/ground application.
}
\details{
This is a helper function for \code{pull_clean_pur}.
}
