% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-helpers.R
\name{help_write_md}
\alias{help_write_md}
\title{Return a \code{meta_data} data frame.}
\usage{
help_write_md(clean_pur_df, pls_percents, pur_out, location, start_date,
  end_date, radius, buffer_area, mtrs_mtr, section_township)
}
\arguments{
\item{clean_pur_df}{A data frame returned by \code{pull_clean_pur} that
includes data for the county of your location (before running
\code{pull_clean_pur}, you can use the
\code{find_location_county} function to figure this out), the time period,
and the active ingredients or chemical classes for which you want to
calculate exposure.}

\item{pls_percents}{A data frame}

\item{pur_out}{A data frame}

\item{location}{A length-one character string. Either a California address
including street name, city, state, and zip code, or a pair of coordinates
in the form "longitude, latitude".}

\item{start_date}{Optional. "yyyy-mm-dd" specifying the start date for
exposure estimation. This date should be present in the \code{clean_pur_df}
data frame.}

\item{end_date}{Optional. "yyyy-mm-dd" specifying the end date for exposure
estimation. This date should be present in the \code{clean_pur_df}
data frame.}

\item{radius}{A numeric value greater than zero that gives the radius in meters
defining the buffer around your location in which you would like to
calculate exposure. For reference, the length and width of a PLS section is
about 1,609 meters (1 mile). That of a township could range from about
9,656 to 11,265 meters (6-7 miles).}

\item{buffer_area}{A numeric value}

\item{mtrs_mtr}{Either \code{MTRS} or \code{MTR}. Not quoted}

\item{section_township}{Either \code{section} or \code{township}. Not quoted}
}
\value{
A data frame with the twelve columns in the
\code{calculate_exposure$meta_data} data frame.
}
\description{
\code{help_write_md} returns a data frame to be output as the \code{meta_data}
element in the list returned from \code{calculate_exposure}.
}
\details{
This is a helper function for \code{help_calculate_exposure}.
}
