% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-map.R
\name{map_counties}
\alias{map_counties}
\title{Map a county's location in California}
\usage{
map_counties(counties_or_df, one_plot = TRUE, fill_color = "red",
  transparency = 0.5)
}
\arguments{
\item{counties_or_df}{A character vector of county names or county codes.
You can use the \code{county_codes} dataset included with this package to
check out PUR county names and codes. This argument can also be a data frame
with a \code{county_cd}, \code{county_name}, or \code{county_code} column. If
you provide a data frame, a plot for every county with data in that dataset
will be output.}

\item{one_plot}{TRUE / FALSE. If you provided multiple counties, whether you
would like county outlines plotted in the same plot (TRUE), or if you would
like separate plots returned in a list (FALSE). The default is TRUE}

\item{fill_color}{A character string giving either a ggplot2 color or a
hex color code ("#0000FF", for example). The default is "red".}

\item{transparency}{A numeric value in [0,1]. Corresponds to the
\code{ggplot2::alpha} transparency argument. Numbers closer to 0 will result
in more transparency. The default is 0.5.}
}
\value{
A ggplot or a list of ggplots of Califnornia with shaded-in counties.
List element names correspond to county names.
}
\description{
\code{map_counties} returns one or multiple plots with county
locations in California given either a vector of county names or codes,
or a PUR data frame with a \code{county_cd}, \code{county_name}, or
\code{county_code} column (A data frame returned from either
\code{pull_pur_file}, \code{pull_raw_pur}, or \code{pull_clean_pur}).
}
\examples{
\dontrun{
map_counties("fresno")

pur_df <- pull_clean_pur(1990, counties = c("01", "05", "12"), verbose = FALSE)
map_counties(pur_df)

plot_list <- map_counties(c("san bernardino", "ventura"), one_plot = TRUE)
names(plot_list)
plot_list[[1]]
plot_list[[2]]
}

}
