% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-plot.R
\name{plot_application_timeseries}
\alias{plot_application_timeseries}
\title{Plot time series}
\usage{
plot_application_timeseries(clean_pur_df)
}
\description{
Plot time series
}
\examples{
\dontrun{
clean_pur <- pull_clean_pur(1990:2015, "fresno", chemicals = "methyl bromide",
                            sum_application = TRUE)
}
}
