% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-plot.R
\name{plot_county_locations}
\alias{plot_county_locations}
\title{Plot a county's location in California.}
\usage{
plot_county_locations(counties_or_df, separate_plots = FALSE,
  fill_color = "red", alpha = 0.5)
}
\arguments{
\item{counties_or_df}{A character vector of county names, pur codes, or fips
codes. You can use the \code{county_codes} dataset included with this
package to check out PUR county names and codes. This argument can also be
a data frame with a \code{county_cd}, \code{county_name}, \code{pur_code},
or \code{fips_code} column. If you provide a data frame, a plot for every
county with data in that dataset will be output.}

\item{separate_plots}{TRUE / FALSE. If you provided multiple counties, whether you
would like county outlines plotted in the same plot (FALSE), or if you would
like separate plots returned in a list (TRUE). The default is FALSE.}

\item{fill_color}{A character string giving either a ggplot2 color or a
hex color code ("#0000FF", for example). The default is "red".}

\item{alpha}{A number in [0,1] specifying the transparency of the fill
color. Numbers closer to 0 will result in more transparency. The default is
0.5.}
}
\value{
A ggplot or a list of ggplots of Califnornia with shaded-in counties.
List element names correspond to county names.
}
\description{
\code{plot_county_locations} returns one or multiple plots with county
locations in California given either a vector of county names or codes,
or a PUR data frame with a \code{county_cd}, \code{county_name},
\code{pur_code}, or \code{fips_code} column (A data frame returned from either
\code{pull_pur_file}, \code{pull_raw_pur}, or \code{pull_clean_pur}).
}
\examples{
\dontrun{
plot_county_locations("fresno")

pur_df <- pull_clean_pur(1990, counties = c("01", "05", "12"))
plot_county_locations(pur_df)

plot_list <- plot_county_locations(c("san bernardino", "ventura"),
                                   separate_plots = TRUE)
names(plot_list)
plot_list[[1]]
plot_list[[2]]
}

}
