% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-plot.R
\name{plot_locations_exposure}
\alias{plot_locations_exposure}
\title{Plot exposure for multiple locations in a county.}
\usage{
plot_locations_exposure(exposure_df, section_township = "section",
  fill_option = "viridis")
}
\arguments{
\item{exposure_df}{A data frame returned from \code{write_exposure} with 10
columns, including \code{exposure}, \code{location}, \code{radius},
\code{longitude}, and \code{latitude}. This data frame should be filtered
so that there is one exposure value per location.}

\item{section_township}{Either "section" (the default) or "township". Specifies
which PLS unit to plot the county by.}

\item{fill_option}{A palette from the colormap package. The default is
"viridis". To see colormap palette options, visit
\url{https://bhaskarvk.github.io/colormap/} or run
\code{colormap::colormaps}.}
}
\value{
A plot with one point per location, colored by each location's
corresponding exposure value.
}
\description{
\code{plot_locations_exposure} returns a plot of exposure to applied pesticides
for multiple locations in a given county.
}
