% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-helpers.R
\name{plot_pls}
\alias{plot_pls}
\title{Return a map for a given exposure estimate.}
\usage{
plot_pls(start_date, end_date, chemicals, aerial_ground, none_recorded,
  data_pls, gradient, location_longitude, location_latitude, buffer_df, buffer2,
  buffer, buffer_or_county, alpha, clean_pur, pls_labels, pls_labels_size,
  percentile, color_by)
}
\arguments{
\item{start_date}{A date}

\item{end_date}{A date}

\item{data_pls}{A data frame}

\item{gradient}{A character vector of hex color codes}

\item{location_longitude}{A numeric longitude value}

\item{location_latitude}{A numeric latitude value}

\item{buffer_df}{A data frame}

\item{buffer2}{A data frame}

\item{buffer}{A gpc.poly object}

\item{buffer_or_county}{"buffer" or "county"}

\item{alpha}{A number in [0,1]}

\item{clean_pur}{A \code{clean_pur_df} data frame}

\item{pls_labels}{TRUE / FALSE}

\item{pls_labels_size}{A number}

\item{percentile}{A numeric vector}

\item{color_by}{"amount" or "percentile"}
}
\description{
For a unique combination of time periods, chemicals, and application methods,
\code{plot_pls} returns a plot showing amounts of pesticides applied for PLS
units intersecting with the buffer.
}
\details{
This is a helper function for \code{map_exposure}.
}
