% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-pull_shapefiles.R
\name{plot_tidy}
\alias{plot_tidy}
\title{Plot data frame spatial objects.}
\usage{
plot_tidy(df)
}
\arguments{
\item{df}{A data frame returned from the \code{spdf_to_df} function.}
}
\value{
A ggplot2 plot of the county.
}
\description{
\code{plot_tidy} plots a data frame spatial object. (A
SpatialPolygonsDataFrame that has been "tidied" using the broom package.)
Meant to be analogous to the ease of using plot() to quickly view a
SpatialPolygonDataFrame object.
}
\examples{
\dontrun{
shp <- pull_spdf("san diego", "township")
df <- spdf_to_df(shp)
plot_tidy(df)
}
}
