% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{products_2000}
\alias{products_2000}
\title{An example PUR product table}
\format{A data frame with 41,076 rows and 5 columns.}
\usage{
products_2000
}
\description{
A PUR Product Table pulled using `pull_product_table` for the year 2000.
}
\keyword{datasets}
