% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_pur.R
\name{pull_pur_file}
\alias{pull_pur_file}
\title{Pull raw PUR file for a year and county or counties.}
\usage{
pull_pur_file(year, county, download_progress = FALSE)
}
\arguments{
\item{county}{A character string giving either a county name or a two digit
county code. Not case sensitive. California names and county codes as they
appear in PUR dataset can be found in the county_codes dataset available
with this package. For example, to return data for Alameda county, enter
either "alameda" or "01" for the county argument.}

\item{download_progress}{TRUE / FALSE indicating whether you would like a
message and progress bar printed for each year of PUR data that is downloaded.
The default value is FALSE.}
}
\value{
A data frame with 33 columns. Year for which data was pulled is
  indicated by \code{applic_dt}; county is indicated by \code{county_cd}.

\dontrun{
raw_file <- pull_pur_file(1999, "ventura")
}
}
\description{
\code{pull_pur_file} pulls the raw PUR dataset for a particular year.
}
