% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-pull.R
\name{pull_raw_pur}
\alias{pull_raw_pur}
\title{Pull raw PUR data by counties and years.}
\usage{
pull_raw_pur(years = "all", counties = "all", verbose = TRUE,
  quiet = FALSE)
}
\arguments{
\item{years}{A four-digit numeric year or vector of years in the range of
1990 to 2015. Indicates the years for which you would like to pull PUR
datasets. \code{years == "all"} will pull data from 1990 through 2015.}

\item{counties}{A vector of character strings giving either a county name,
two digit PUR county codes, or six-digit FIPS county codes for each county.
Not case sensitive. California names, county codes as they appear in PUR
datasets, and FIPS county codes can be found in the \code{county_codes}
dataset available with this package. For example, to return data for
Alameda county, enter either "alameda", "01", or "06001" for the
\code{counties} argument. \code{counties = "all"} will return data for all
58 California counties (this will take a while to run).}

\item{verbose}{TRUE / FALSE indicating whether you would like a single message
printed indicating which counties and years you are pulling data for. The
default value is TRUE.}

\item{quiet}{TRUE / FALSE indicating whether you would like a
message and progress bar printed for each year of PUR data that
is downloaded. The default value is FALSE.}
}
\value{
A data frame with 33 columns. Different years and counties for which
  data was pulled are indicated by \code{applic_dt} and \code{county_cd},
  respectively.
}
\description{
\code{pull_raw_pur} pulls a raw PUR dataset for a given year and vector of
  California counties.
}
\section{Note}{

 \itemize{
   \item{For documentation of raw PUR data, see the Pesticide Use
  Report Data User Guide & Documentation document published by the California
  Department of Pesticide Regulation. This file is saved as "cd_doc.pdf" in any
  "pur[year].zip" file between 1990 and 2015 found here:
  \url{ftp://transfer.cdpr.ca.gov/pub/outgoing/pur_archives/}.}
  \item{If this function returns an error (because the FTP site is down, for
  example), check your working directory. You may need to change it back from
  a temporary directory.}
}
}

\examples{
\dontrun{
df <- pull_raw_pur(years = c(2000, 2010), counties = c("butte", "15", "06001"))
df2 <- pull_raw_pur(years = 2015, counties = c("colusa"))
}
}
