% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-pull_shapefiles.R
\name{pull_spdf}
\alias{pull_spdf}
\title{Pull California county SpatialPolygonsDataFrame}
\usage{
pull_spdf(county, section_township = "section", download_progress = FALSE)
}
\arguments{
\item{county}{A vector of character strings giving either a county names or
two digit PUR county codes. Not case sensitive. California names and county
codes as they appear in PUR datasets can be found in the \code{county_codes}
dataset available with this package.}

\item{section_township}{Either "section" (the default) or "township".
Specifies whether you would like to pull a section- or township-level
SpatialPolygonsDataFrame.}

\item{download_progress}{TRUE / FALSE indicating whether you would like a
message and progress bar printed for the shapefile that is downloaded.
The default value is FALSE.}
}
\value{
A SpatialPolygonsDataFrame object.
}
\description{
\code{pull_spdf} pulls either the section or township-level
SpatialPolygonsDataFrame from a county's Geographic Information System (GIS)
shapefile.
}
\section{Source}{

SpatialPolygonDataFrame objects are downloaded from GIS shapefiles provided
by the California Department of Pesticide Regulation:
\url{http://www.cdpr.ca.gov/docs/emon/grndwtr/gis_shapefiles.htm}
}

\examples{
\dontrun{
trinity_shp <- pull_spdf("trinity", download_progress = TRUE)
plot(trinity_shp)

del_norte_shp <- pull_spdf("08", "township", download_progress = TRUE)
plot(del_norte_shp)
}
}
