% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_pur.R
\name{pur_data}
\alias{pur_data}
\title{Pull PUR data by year and vector of chemicals.}
\usage{
pur_data(county, years, chemicals, sum = TRUE, unit = "section")
}
\arguments{
\item{county}{Either a two-digit}

\item{years}{A four-digit numeric year in the range of 1990 to 2015. Indicates
the year in which you would like to match chemical codes.
\code{years == "all"} will pull data from 1990 through 2015.}

\item{sum}{TRUE / FALSE indicating if you would like to sum the amounts of
applied active ingredients by day and by the geographic unit given in
\code{unit}. The default value is TRUE.}

\item{unit}{A character string giving either "section" or "township".
Specifies whether applications of each active ingredient should be summed
by California section (the default, \code{MTRS}) or by township
(\code{MTR}).}
}
\value{
A data frame with twelve columns:
  \itemize{
    \item \code{chem_code} and \code{chemname} correspond to chemicals given
    in the \code{chemicals} argument.
    \item \code{lbs_chm_used} gives the amount of pesticide, in pounds,
    applied for a given \code{date} and active ingredient.
    \item \code{MTRS} and \code{MTR} indicate the section and township where
    application of a given active ingredient on a given day took place,
    respectively,
    \item \code{county_name} gives the county of application,
    \item \code{date} the date of application, and
    \item \code{use_no} gives an ID identifing unique application of an
    active ingredient.
    \item The \code{outlier} column is a logical value indicating whether the
    amount listed in \code{lbs_chm_used} has been corrected large amounts
    entered in error. The algorithm for identifying and replacing outliers
    was developed based on methods used by Gunier et al. (2001). Please see
    the package vignette for more detail regarding these methods.
  }
}
\description{
\code{pur_data} returns a data frame of Pesticide Use Report data filtered
by county, year, active ingredients, and summed by either section or
township.
}
\examples{
\dontrun{
df <- pur_data(years = 2000, chemicals = "methyl bromide")
df2 <- pur_data(years = 1990:1995, chemicals = c("methyl bromide",
                "trifluralin"), unit = "township")
df3 <- pur_data(years = 2010:2012, chemicals = "ammonia", sum = FALSE)
}
}
