% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-helpers.R
\name{pur_filt_df}
\alias{pur_filt_df}
\title{Find PLS units intersecting with a buffer.}
\usage{
pur_filt_df(pls, pls_quote)
}
\arguments{
\item{pls}{Either \code{MTRS} (sections) or \code{MTR} (townships). Not quoted.}

\item{pls_quote}{Either \code{"MTRS"} or \code{"MTR"}}
}
\value{
A list with four elements:
\describe{
  \item{pur_filt}{A cleaned PUR data frame filtered to PLS units intersecting
  with a buffer.}
  \item{comb_df_filt}{A spatial data frame with intersecting PLS units and
  a buffer.}
  \item{pls_intersections}{A data frame with two columns: \code{pls} and
  \code{percent}, the corresponding percent intersection with the buffer.}
  \item{pls_int}{A character vector with all PLS units intersecting with the
  buffer.}
}
}
\description{
\code{pur_filt_df} filters a SpatialPolygonsDataFrame to include only PLS
units intersecting with a buffer, and filters the data frame returned from
\code{pull_clean_pur} to include only those sections or townships.
}
\details{
This is a helper function for \code{calculate_exposure}.
}
