% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-helpers.R
\name{pur_out_df}
\alias{pur_out_df}
\title{Find summed applied active ingredients}
\usage{
pur_out_df(...)
}
\arguments{
\item{...}{A list of variables to group by. Options include \code{section},
\code{township}, \code{chemical_class}, and \code{aerial_ground}. Not quoted.}
}
\value{
A data frame a \code{kg} column and one to three additional columns,
depending on the grouping variables.
}
\description{
\code{pur_out_df} finds the summed amount of applied active ingredients by
section or township, chemical class, and aerial/ground application.
}
\details{
This is a helper function for \code{daterange_calcexp}.
}
