% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purexposure.R
\docType{package}
\name{purexposure}
\alias{purexposure}
\alias{purexposure-package}
\title{purexposure: A package for working with CA Pesticide Use Registry data.}
\description{
The \code{purexposure} package provides functions to pull data from California's
Pesticide Use Registry (PUR), as well as to calculate exposure to and
visualize active ingredients present in applied pesticides. The main function
categories are \code{find_*}, \code{pull_*}, \code{calculate_*}, and
\code{plot_*}. These are the most important functions from each category:
}
\section{find_* functions}{

\code{find_} functions help with searches of PUR chemical, county, and
product codes.
\itemize{
  \item \code{find_chemical_codes}: Pull active ingredient chemical codes
  from PUR Chemical Lookup Tables.
  \item \code{find_counties}: Find California county codes or names.
  \item \code{find_product_name}: Find Pesticide Product names and registration
  numbers from PUR Product Lookup Tables.
}
}

\section{pull_* functions}{

\code{pull_} functions facilitate downloading data from the CA Department
of Pesticide Regulation's website.
\itemize{
  \item \code{pull_raw_pur}: Pull raw PUR data by counties and years.
  \item \code{pull_clean_pur}: Pull cleaned PUR data by counties, years, and
  active ingredients.
}
}

\section{calculate_* function}{

The \code{calculate_exposure} function calculates exposure (in kg/m^2) to
applied pesticides for a given location, buffer extending from that location,
time period, and active ingredients.
}

\section{plot_* functions}{

\code{plot_} functions help with visualizations of application.
\itemize{
  \item \code{plot_county_application}: Plot pesticide application by county,
  summed by section or township.
  \item \code{plot_exposure}: Plot exposure to applied pesticides at a
  particular location.
  \item \code{plot_application_timeseries}: Plot time series of active
  ingredients in applied pesticides.
}
}

