% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_pur.R
\name{raw_pur}
\alias{raw_pur}
\title{Pull raw PUR data by county and year}
\usage{
raw_pur(county, year)
}
\arguments{
\item{county}{A character string giving either a county name or a two digit
county code. California names and county codes as they appear in PUR
dataset can be found in the \code{purexposure::county_codes} data frame.
For example, to return data for Alameda county, enter either "alameda" or
"01" for the \code{county} argument.}

\item{year}{A four-digit numeric year in the range of 1990 to 2015. Indicates
the year in which you would like to search for active ingredients.}
}
\value{
A data frame with
}
\description{
\code{raw_pur} pulls a raw PUR dataset for a given California county and
year.
}
\examples{
\dontrun{

}
}
