% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-helpers.R
\name{remove_cols}
\alias{remove_cols}
\title{Remove columns with all missing values}
\usage{
remove_cols(col_quote, df)
}
\arguments{
\item{col_quote}{A quoted column name}

\item{df}{A data frame}
}
\value{
A data frame with two columns: \code{col} gives the column name, and
\code{all_missing} is a logical value.
}
\description{
Given a quoted column name and its data frame, \code{remove_cols} determines
if that column has all missing values or not.
}
\details{
This is a helper function for \code{sum_application}.
}
