% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-helpers.R
\name{scale_fill_gradientn2}
\alias{scale_fill_gradientn2}
\title{Include alpha in ggplot2::scale_fill_gradientn()}
\usage{
scale_fill_gradientn2(..., colours, values = NULL, space = "Lab",
  na.value = "grey50", guide = "colourbar", colors, alpha = NULL)
}
\description{
This function adds an `alpha` argument to scale_fill_gradientn() from the
ggplot2 package.
}
