% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-helpers.R
\name{single_county_code}
\alias{single_county_code}
\title{Find California county code or name}
\usage{
single_county_code(county, find = "codes")
}
\arguments{
\item{county}{A character string giving either a county name or
two digit PUR county code. Not case sensitive. California names and county
codes as they appear in PUR datasets can be found in the \code{county_codes}
dataset available with this package.}

\item{return}{Either "codes" to return county code (the default) or "names"
to return county name.}
}
\value{
If \code{return = "codes"}, a two-character string giving
  the corresponding PUR county codes. If \code{return = "names"}, a county
  name.
}
\description{
Given a county, \code{single_county_code} returns either the PUR
county code or name.
}
\details{
This is a helper function for \code{find_counties}.
}
\examples{
single_county_code("01", find = "names")
single_county_code("contra costa", find = "codes")
}
