% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-helpers.R
\name{spdf_to_df}
\alias{spdf_to_df}
\title{Convert county SpatialPolygonsDataFrame to a tidy data frame.}
\usage{
spdf_to_df(spdf)
}
\arguments{
\item{spdf}{A SpatialPolygonsDataFrame object returned from
the \code{pull_spdf} function.}
}
\value{
A data frame with 24 columns if the \code{spdf} object is on the
section level and 23 columns if the \code{spdf} object is on the township
level.
}
\description{
\code{spdf_to_df} converts a SpatialPolygonsDataFrame object returned from
the \code{pull_spdf} function to a data frame.
}
\examples{
\dontrun{
df <- spdf_to_df(pull_spdf("fresno"))
df2 <- spdf_to_df(pull_spdf("sonoma"))

# use dataframe_plot() function to easily plot the output data frames:
dataframe_plot(df)
dataframe_plot(df2)
}
}
