% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-helpers.R
\name{sum_application}
\alias{sum_application}
\title{Sum application}
\usage{
sum_application(df, sum, unit, aerial_ground, ...)
}
\arguments{
\item{df}{A data frame from \code{pull_clean_pur} before summing has taken
place}

\item{sum}{"all" or "chemical_class"}

\item{unit}{either "section" or "township"}

\item{aerial_ground}{TRUE / FALSE}

\item{...}{grouping variables}
}
\value{
A data frame with 12 columns. Some will have missing values.
}
\description{
\code{sum_application} sums application of a PUR dataset by chemicals,
PLS unit, and aerial/ground application.
}
\details{
This is a helper function for \code{pull_clean_pur}.
}
