% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_pur.R
\name{tibble_to_vector}
\alias{tibble_to_vector}
\title{Extract a character vector from a tibble with one column.}
\usage{
tibble_to_vector(tib)
}
\arguments{
\item{df}{A tibble with one column.}
}
\value{
A character vector.
}
\description{
\code{tibble_to_vector} extracts the values from a tibble with one column and
saves them as a character vector.
}
\examples{
tibble_to_vector(tibble::tibble(x = 1:3))
}
