% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-shapefiles.R
\name{view_county_df}
\alias{view_county_df}
\title{Plot data frame spatial objects.}
\usage{
view_county_df(df)
}
\arguments{
\item{df}{A data frame returned from the \code{spdf_to_df} function.}
}
\value{
A ggplot2 plot of the county.
}
\description{
\code{view_county_df} plots a data frame spatial object. Meant to be
analogous to the ease of using plot() to quickly view a
SpatialPolygonDataFrame object.
}
\examples{
\dontrun{
shp <- pull_spdf("san diego", "township")
df <- spdf_to_df(shp)
view_county_df(df)
}
}
