% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-calculate.R
\name{calculate_exposure}
\alias{calculate_exposure}
\title{Calculate exposure to active ingredients present in applied pesticides.}
\usage{
calculate_exposure(clean_pur_df, location, radius, time_period = NULL,
  start_date = NULL, end_date = NULL, chemicals = "all")
}
\arguments{
\item{clean_pur_df}{A data frame returned by \code{clean_pur_data} that
includes data relevant to the county in which your location is located (you
can use the \code{find_location_county} function to figure this out), the
time period, and the active ingredients or chemical classes for which you
want to calculate exposure.}

\item{location}{A character string. Either a California address inclding
street name, city, state, and zip code, or a pair of coordinates in the
form "longitude, latitude".}

\item{radius}{A numeric value greater than zero that gives the radius in meters
defining the buffer around your location in which you would like to
calculate exposure. For reference, the length and width of a PLS section is
about 1,609 meters (1 mile). That of a township is about 22,249 meters (48
miles).}

\item{time_period}{Optional. A character string giving a time period over which you
would like to calculate exposure. For example, if you enter "6 months" for
\code{time_period}, \code{calculate_exposure} will calculate exposure for
every six month period starting from the earliest date present in the
\code{clean_pur_df} data frame. Start and end dates can be optionally specified
with the \code{start_date} and \code{end_date} arguments. Alternatively, to
calculate exposure over only one time period, you can leave this argument
NULL and specify start and end dates.}

\item{start_date}{Optional. "yyyy-mm-dd" specifying the start date for
exposure estimation. This date should be present in the \code{clean_pur_df}
data frame.}

\item{end_date}{Optional. "yyyy-mm-dd" specifying the end date for exposure
estimation. This date should be present in the \code{clean_pur_df}
data frame.}

\item{chemicals}{A character string indicating active ingredients for which you
would like to calculate exposure. The default is "all", which will calculate
exposure to the summed active ingredients present in the \code{clean_pur_df}
data frame. You can also enter a particular active ingredient that's present
in your \code{clean_pur_df} data frame, or "chemical_class" to calculate
exposure to each of the chemical classes present in the \code{chemical_class}
column of your \code{clean_pur_df} data frame.}
}
\value{
A data frame with ..
}
\description{
For a particular location, buffer radius from that location, date range, and
active ingredient or class of active ingredients, \code{calculate_exposure}
returns an estimate of exposure in kg of active ingredient per m^2.
}
\section{Note}{

 \itemize{
   \item{If the \code{time_period}, \code{start_date}, and \code{end_date}
         arguments are all left as NULL (their defaults), then exposure will
         be estiamted across the entire date range of the \code{clean_pur_df}
         data frame.}
   \item{If you pulled PUR data from \code{clean_pur_data} specifying
         \code{sum_application = TRUE} and \code{unit = "township"}, then
         exposure will be calculated based on townships. Because PLS townships
         are geographically coarse, this is not ideal- making sure that you
         have a "sections" column in your \code{pur_data} data frame (either
         don't sum application or leave \code{unit} as \code{"section"}) will
         give better exposure estimations.}
}
}

\examples{
\dontrun{

pur_data <- clean_pur_data()

# Time periods

# Time periods with specified start and end date

# Single time period

# All active ingredients

# A single active ingredient

# By chemical class
pur_data_classes <- clean_pur_data()

}
}
