% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-calculate.R
\name{calculate_exposure_bydate}
\alias{calculate_exposure_bydate}
\title{Calculate exposure for multiple start and end dates}
\usage{
calculate_exposure_bydate(start_date, end_date)
}
\arguments{
\item{start_date}{"yyyy-mm-dd"}

\item{end_date}{"yyyy-mm-dd"}
}
\value{
A data frame with two columns. \code{row_out} is a nested data frame
with one row and seven columns. \code{meta_data} is a nested list with one
element: a data frame with as many rows as there are sections or townships
intersecting with the specified buffer and 11 columns.
}
\description{
\code{calculate_exposure_bydate} takes a start and end date and returns
exposure information. Written to be used in the \code{calculate_exposure}
function- has dependencies for other variables (pur_filt data frame, radius,
location, and chemicals) produced there.
}
