% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-helpers.R
\name{daterange_calcexp}
\alias{daterange_calcexp}
\title{Return exposure data for a single start and end date}
\usage{
daterange_calcexp(start_date, end_date, aerial_ground, chemicals, clean_pur_df,
  location, pls_percents, pur_filt, radius)
}
\arguments{
\item{start_date}{A date, "yyyy-mm-dd"}

\item{end_date}{A date, "yyyy-mm-dd"}

\item{aerial_ground}{TRUE / FALSE}

\item{chemicals}{Either "all" or "chemical_class"}
}
\value{
A nested data frame with two columns: The \code{row_out} column
contains the \code{exposure} data frame for the date range, and
\code{meta_data} contains the \code{meta_data} data frame for the date range.
}
\description{
For a single date range, \code{daterange_calcexp} returns the
\code{exposure} and \code{meta_data} data frames to be output in the
\code{calculate_exposure} list as a nested data frame.
}
\details{
This is a helper function for \code{calculate_exposure}.
}
