% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-misc.R
\name{df_plot}
\alias{df_plot}
\title{Plot data frame spatial objects.}
\usage{
df_plot(df)
}
\arguments{
\item{df}{A data frame returned from the \code{spdf_to_df} function.}
}
\value{
A ggplot2 plot of the county.
}
\description{
\code{df_plot} plots a data frame spatial object. (A
SpatialPolygonsDataFrame that has been "tidied" using the broom package.)
Meant to be analogous to the ease of using plot() to quickly view a
SpatialPolygonDataFrame object.
}
\examples{
library(magrittr)
\donttest{
fresno <- readRDS(system.file("extdata", "fresno_spdf.rds", package = "purexposure"))
fresno \%>\% spdf_to_df \%>\% df_plot()}
\donttest{
pull_spdf("fresno") \%>\% spdf_to_df() \%>\% df_plot()
pull_spdf("san diego", "township") \%>\%
   spdf_to_df() \%>\%
   df_plot()
}
}
