% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-helpers.R
\name{exp_df}
\alias{exp_df}
\title{Return a the \code{meta_data} data frame}
\usage{
exp_df(clean_pur_df, pls_percents, pls, pur_out, location, start_date, end_date,
  radius, buffer_area, mtrs_mtr, section_township)
}
\arguments{
\item{buffer_area}{}

\item{mtrs_mtr}{Either \code{MTRS} or \code{MTR}. Not quoted.}

\item{section_township}{Either \code{section} or \code{township}. Not quoted.}
}
\value{
A data frame with the twelve columns in the
\code{calculate_exposure$meta_data} data frame.
}
\description{
\code{exp_df} returns a data frame to be output as the \code{meta_data}
element in the list returned from \code{calculate_exposure}.
}
\details{
This is a helper function for \code{daterange_calcexp}.
}
