% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-find.R
\name{find_location_county}
\alias{find_location_county}
\title{Find the county from an address or coordinate pair.}
\usage{
find_location_county(location, return = "name")
}
\arguments{
\item{location}{A character string. Either a California address including
street name, city, state, and zip code, or a pair of coordinates in the
form "longitude, latitude".}

\item{return}{Either "name" to return county name (the default) or "code"
to return county code.}
}
\value{
A character string giving the California county where the address or
coordinate pair given in \code{location} is located.
}
\description{
Given a California address or longitude/latitude coordinates,
\code{find_location_county} returns the corresponding California county or
PUR code.
}
\examples{
\dontrun{
address <- "13883 Lassen Ave, Helm, CA 93627"
find_location_county(location = address)

long_lat <- c("-120.09789, 36.53379")
find_location_county(location = long_lat)
}
}
