% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-find.R
\name{find_product_name}
\alias{find_product_name}
\title{Find pesticide product names and registration numbers from PUR Product Lookup
Tables.}
\usage{
find_product_name(years, products = "all", quiet = FALSE, by_year = FALSE,
  ...)
}
\arguments{
\item{years}{A vector of four digit years in the range of 1990 to 2015.}

\item{products}{A character string or a vector of character strings with
pesticide product names that you would like to search for. Not case
sensitive. The default is "all", which will return all pesticide products
applied for a given year.}

\item{quiet}{TRUE / FALSE indicating whether you would like a
message and progress bar printed for the product table that is downloaded.
The default value is FALSE.}

\item{by_year}{TRUE / FALSE for whether you would like PUR Product Lookup
Tables separated by year (in a `year` column). If `by_year` is `FALSE`, the
default, a data frame is returned with unique results from all years given
in the `years` argument.}

\item{...}{Used internally.}
}
\value{
A data frame with seven columns:
\describe{
  \item{prodno}{The CA registration number. Can be matched with the
  \code{prodno} in a raw or cleaned PUR data set.}
  \item{prodstat_ind}{Character. An indication of product registration status:
  \itemize{
  \item A = Active
  \item B = Inactive
  \item C = Inactive, Not Renewed
  \item D = Inactive, Voluntary Cancellation
  \item E = Inactive, Cancellation
  \item F = Inactive, Suspended
  \item G = Inactive, Invalid Data
  \item H = Active, Suspended}}
  \item{product_name}{Character. The name of the product taken from the
  registered product label. May have been modified by DPR's Registration Branch
  to ensure uniqueness.}
  \item{signlwrd_ind}{Integer. The signal word printed on the front of the
product label:
  \itemize{
  \item 1 = Danger (Poison)
  \item 2 = Danger (Only)
  \item 3 = Warning
  \item 4 = Caution
  \item 5 = None}}
  \item{product}{Product name search terms.}
  \item{year}{The year for which product table information was pulled.
  Included if `by_year` is set to TRUE.}
}
}
\description{
For a vector of years and product search terms, \code{find_product_name}
returns a data frame with corresponding product registration numbers,
\code{prodno}, indicator codes, and product names.
}
\details{
Product tables are pulled by year from the CDPR's FTP server. Downloaded
tables are saved in a temporary environment, which is deleted at the end of
the current R session.
}
\examples{
\dontshow{
product_df <- readRDS(system.file("extdata", "products_2000.rds",
                                  package = "purexposure"))
prod_df <- find_product_name(2000, "mosquito", product_df = product_df)}
\donttest{
prod_df <- find_product_name(2000, "mosquito")
prod_df2 <- find_product_name(2010, c("insecticide", "rodenticide"))
}
}
