% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-map.R
\name{map_exposure}
\alias{map_exposure}
\title{Map exposure to pesticides by PLS unit and buffer.}
\usage{
map_exposure(exposure_list)
}
\description{
- list from calculate_exposure() for single location and radius
- this function returns a plot for each date range, aerial/ground,
and chemicals (if chemical_class was used in calculate_exposure()
combination)
- amounts of active ingredients are plotted based on the % intersection of the
PLS unit with the buffer around a location (for time period, a/g, and chemicals)
}
\details{
$ list w/in out_list
$ plot
$ one-row data frame with location, longitude, latidue, exposure,
radius, buffer_area, chemicals, start_date, end_date, aerial_ground
- each one-row data frame and represents a unique exposure estimate for unique
combination of column values
- better way to organize this?
}
