% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-plot.R
\name{plot_application_timeseries}
\alias{plot_application_timeseries}
\title{Plot time series of active ingredients in applied pesticides.}
\usage{
plot_application_timeseries(clean_pur_df, facet = FALSE, axes = "fixed")
}
\arguments{
\item{clean_pur_df}{A data frame returned from \code{pull_clean_pur}.}

\item{facet}{TRUE / FALSE for whether you would like time series
plots to be faceted by unique \code{chemname} or \code{chemical_class}
column values. If \code{facet = FALSE} (the default), all active ingredients
present in the data set will be summed per day.}

\item{axes}{A character string passed on to the \code{scales} argument of
\code{ggplot2::facet_wrap} (\code{"fixed"}, \code{"free"}, \code{"free_x"},
or \code{"free_y"}). The default is \code{"fixed"}.}
}
\value{
A \code{ggplot2} object.
}
\description{
\code{plot_application_timeseries} returns a \code{ggplot2} time series plot
of pesticides present in a \code{pull_clean_pur} data frame. You can choose
whether to facet the time series by active ingredient (\code{chemname}) or by
\code{chemical_class}.
}
\examples{
library(magrittr)
readRDS(system.file("extdata", "fresno_clean.rds", package = "purexposure")) \%>\%
    plot_application_timeseries()
\donttest{
pull_clean_pur(1990:1992, "fresno") \%>\%
    dplyr::filter(chemname \%in\% toupper(c("methyl bromide", "sulfur"))) \%>\%
    plot_application_timeseries(facet = TRUE)
}
}
