% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-plot.R
\name{plot_locations_exposure}
\alias{plot_locations_exposure}
\title{Plot exposure for multiple locations in a county.}
\usage{
plot_locations_exposure(exposure_df, section_township = "section",
  fill = "viridis", alpha = 1)
}
\arguments{
\item{exposure_df}{A data frame returned from \code{write_exposure} with 10
columns, including \code{exposure}, \code{location}, \code{radius},
\code{longitude}, and \code{latitude}. This data frame should be filtered
so that there is one exposure value per location.}

\item{section_township}{Either "section" (the default) or "township". Specifies
which PLS unit to plot the county by.}

\item{fill}{A palette from the colormap package. The default is
"viridis". To see colormap palette options, visit
\url{https://bhaskarvk.github.io/colormap/} or run
\code{colormap::colormaps}.}

\item{alpha}{A number in [0,1] specifying the transparency of fill colors.
Numbers closer to 0 will result in more transparency. The default is 1.}
}
\value{
A plot with one point per location, colored by each location's
corresponding exposure value.
}
\description{
\code{plot_locations_exposure} returns a plot of exposure to applied pesticides
for multiple locations in a given county.
}
\examples{
\dontrun{
fresno <- pull_clean_pur(2000, "fresno")
df <- data.frame(location = c("295 West Saginaw Ave., Caruthers, CA 93609",
                              "55190 Point Rd., Big Creek, CA 93605"),
                 start_date = "2000-01-01", end_date = "2000-12-31")
write_exposure(fresno, df, 3000, "~/Documents/fresno")
}

}
