% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-pull.R
\name{pull_pur_file}
\alias{pull_pur_file}
\title{Pull raw PUR file for a single year and a county or counties.}
\usage{
pull_pur_file(year, counties = "all", download_progress = FALSE)
}
\arguments{
\item{counties}{A character vector giving either county names or two digit
county codes. Not case sensitive. California names and county codes as they
appear in PUR datasets can be found in the county_codes dataset available
with this package. For example, to return data for Alameda county, enter
either "alameda" or "01" for the county argument. \code{counties = "all"}
will pull data for all 58 California counties.}

\item{download_progress}{TRUE / FALSE indicating whether you would like a
message and progress bar printed for each year of PUR data that is downloaded.
The default value is FALSE.}
}
\value{
A data frame with 33 columns. Counties are indicated by
  \code{county_cd}; the year for which data was pulled is indicated by
  \code{applic_dt}.
}
\description{
\code{pull_pur_file} pulls the raw PUR dataset for a particular year and
saves datasets for specified counties in a data frame.
}
\section{Note}{

If this function returns an error (because the FTP site is down, for
example), check your working directory. You may have to change it back from a
temporary directory.

\dontrun{
raw_file <- pull_pur_file(1999, c("40", "ventura", "yuba"))
raw_file <- pull_pur_file(2015, "all")
}
}

