% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_pur.R
\name{pull_raw_pur}
\alias{pull_raw_pur}
\title{Pull raw PUR data by counties and years}
\usage{
pull_raw_pur(years, counties, verbose = TRUE, download_progress = FALSE)
}
\arguments{
\item{years}{A four-digit numeric year or vector of years in the range of
1990 to 2015. Indicates the years for which you would like to pull PUR
datasets. \code{years == "all"} will pull data from 1990 through 2015.}

\item{counties}{A character string or vector of character strings giving either
county names or a two digit county code for each county. Not case sensitive.
California names and county codes as they appear in PUR dataset can be found
in the \code{county_codes} dataset available with this package. For example,
to return data for Alameda county, enter either "alameda" or "01" for the
\code{counties} argument.}

\item{verbose}{TRUE / FALSE indicating whether you would like a single message
printed indicating which counties and years you are pulling data for. The
default value is TRUE.}

\item{download_progress}{TRUE / FALSE indicating whether you would like a
message and progress bar printed for each year of PUR data that is downloaded.
The default value is FALSE.}
}
\value{
A data frame with 33 columns. Different years and counties for which
  data was pulled are indicated by \code{applic_dt} and \code{county_cd},
  respectively.
}
\description{
\code{pull_raw_pur} pulls a raw PUR dataset for a given California county and
year.
}
\section{Note}{
 For documentation of raw PUR data, see the Pesticide Use
  Report Data User Guide & Documentation document published by the California
  Department of Pesticide Regulation. This file is saved as "cd_doc.pdf" in any
  "pur[year].zip" file between 1990 and 2015 found here:
  \url{ftp://transfer.cdpr.ca.gov/pub/outgoing/pur_archives/}.
}

\examples{
\dontrun{
df <- pull_raw_pur(years = 1990:1993, counties = c("01", "02", "10"))
df2 <- pull_raw_pur(years = c(2000, 2010), counties = c("butte", "15", "01"))
df3 <- pull_raw_pur(years = 2015, counties = c("colusa"))
}
}
