% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-pull.R
\name{pull_spdf}
\alias{pull_spdf}
\title{Pull California county SpatialPolygonsDataFrame.}
\usage{
pull_spdf(county, section_township = "section", quiet = FALSE)
}
\arguments{
\item{county}{A character string giving either a county name, two digit PUR
county code, or six-digit FIPS county code. Not case sensitive. California
names and county codes as they appear in PUR data sets can be found in the
\code{county_codes} data set available with this package.}

\item{section_township}{Either "section" (the default) or "township".
Specifies whether you would like to pull a section- or township-level
SpatialPolygonsDataFrame.}

\item{quiet}{TRUE / FALSE indicating whether you would like a
message and progress bar printed for the shapefile that is downloaded.
The default value is FALSE.}
}
\value{
A SpatialPolygonsDataFrame object.
}
\description{
\code{pull_spdf} pulls either the section or township-level
SpatialPolygonsDataFrame from a county's Geographic Information System (GIS)
shapefile.
}
\details{
SpatialPolygonsDataFrame objects are pulled by county from the CDPR's FTP
server. Downloaded SpatialPolygonsDataFrame objects are saved in a temporary
environment, which is deleted at the end of the current R session.
}
\section{Source}{

SpatialPolygonDataFrame objects are downloaded from GIS shapefiles provided
by the California Department of Pesticide Regulation:
\url{http://www.cdpr.ca.gov/docs/emon/grndwtr/gis_shapefiles.htm}
}

\section{Note}{

If this function returns an error (because the FTP site is down, for
example), check your working directory. You may want to change it back from
a temporary directory.
}

\examples{
\donttest{
fresno_shp <- pull_spdf("fresno")
fresno_shp \%>\% spdf_to_df() \%>\% df_plot()
del_norte_shp <- pull_spdf("08", "township")
del_norte_shp \%>\% spdf_to_df() \%>\% df_plot()
}
}
