% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_pur.R
\name{pur_data}
\alias{pur_data}
\title{Pull cleaned PUR data by counties, years, and chemicals}
\usage{
pur_data(years, counties, chemicals, sum_application = TRUE,
  unit = "section", include_ag = TRUE, verbose = TRUE,
  download_progress = FALSE)
}
\arguments{
\item{years}{A four-digit numeric year or vector of years in the range of
1990 to 2015. Indicates the years for which you would like to pull PUR
datasets. \code{years == "all"} will pull data from 1990 through 2015.}

\item{counties}{A character string or vector of character strings giving either
county names or a two digit county code for each county. Not case sensitive.
California names and county codes as they appear in PUR dataset can be found
in the \code{county_codes} dataset available with this package. For example,
to return data for Alameda county, enter either "alameda" or "01" for the
\code{counties} argument.}

\item{chemicals}{A string or vector of strings giving search terms of
chemicals to match with active ingredients present in pesticides applied
in the given year. The default value is "all", which returns codes for all
active ingredients applied in a given year.}

\item{sum_application}{TRUE / FALSE indicating if you would like to sum the
amounts of applied active ingredients by day and by the geographic unit
given in \code{unit}. The default value is TRUE.}

\item{unit}{A character string giving either "section" or "township".
Specifies whether applications of each active ingredient should be summed
by California section (the default) or by township if \code{sum_application} is
\code{TRUE}.}

\item{include_ag}{TRUE / FALSE indicating if you would like to retain
aerial/ground application data when summing application across sections or
townships. The default is TRUE. \code{include_ag = TRUE} could result in
more than one record per section, date, and active ingredient if
\code{aerial_ground} application differs for the same chemical in the same
section and on the same day.}

\item{verbose}{TRUE / FALSE indicating whether you would like a single message
printed indicating which counties and years you are pulling data for. The
default value is TRUE.}

\item{download_progress}{TRUE / FALSE indicating whether you would like a
message and progress bar printed for each year of PUR data that is downloaded.
The default value is FALSE.}
}
\value{
A data frame with 11 columns:
  \itemize{
    \item \code{chem_code} An integer value giving the PUR chemical code
    for the active ingredient applied.
    \item \code{chemname} A character string giving PUR chemical active
    ingredient names. Unique values of \code{chemname} are matched with terms
    provided in the \code{chemicals} argument.
    \item \code{kg_chm_used} A numeric value giving the amount of the active
    ingredient applied (kilograms).
    \item \code{section} A string nine characters long indicating the section
    of application. PLS sections are uniquely identified by a combination of
    base line meridian (S, M, or H), township (01-48), township direction
    (N or S), range (01-47), range direction (E or W) and section number
    (01-36). (This column is not included if \code{sum_application} = TRUE).
    \item \code{township} A string 7 characters long indicating the township
    of application. PLS townships are uniquely identified by a combination of
    base line meridian (S, M, or H), township (01-48), township direction
    (N or S), range (01-47), and range direction (E or W).
    \item \code{county_name} A character string giving the county name where
    application took place.
    \item \code{county_code} A string two characters long giving the PUR county
    code where application took place.
    \item \code{date} the date of application (yyyy-mm-dd).
    \item \code{aerial_ground} A character giving the application method.
    "A" = aerial, "G" = ground, and "O" = other. (This column is not included if
    you \code{sum_application} = TRUE and \code{include_ag} = FALSE.)
    \item \code{use_no} A character ID ID identifing unique application of an
    active ingredient across years. This value is a combination of the raw PUR
    \code{use_no} and the year of application.
    \item \code{outlier} A logical value indicating whether the
    amount listed in \code{lbs_chm_used} has been corrected large amounts
    entered in error. The algorithm for identifying and replacing outliers
    was developed based on methods used by Gunier et al. (2001). Please see
    the package vignette for more detail regarding these methods. (This column
    is not included if \code{sum_application} = TRUE).
  }
}
\description{
\code{pur_data} returns a data frame of cleaned Pesticide Use Report data
filtered by counties, years, and active ingredients. Active ingredients
present in applied pesticides can be summed by either Public Land Survey
(PLS) section or township.
}
\section{Note}{
 For documentation of raw PUR data, see the Pesticide Use
  Report Data User Guide & Documentation document published by the California
  Department of Pesticide Regulation. This file is saved as "cd_doc.pdf" in any
  "pur[year].zip" file between 1990 and 2015 found here:
  \url{ftp://transfer.cdpr.ca.gov/pub/outgoing/pur_archives/}.
}

\examples{
\dontrun{

}
}
