% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purexposure.R
\docType{package}
\name{purexposure}
\alias{purexposure}
\alias{purexposure-package}
\title{purexposure: A package for working with CA Pesticide Use Registry data.}
\description{
The purexposure package provides functions to pull data from California's
Pesticide Use Registry (PUR), as well as to calculate exposure to and
visualize active ingredients present in applied pesticides.
}
