% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-helpers.R
\name{row_out_df}
\alias{row_out_df}
\title{Return a data frame with exposure values and other related data.}
\usage{
row_out_df(start_date, end_date, location, radius, exp, buffer_area, ...)
}
\arguments{
\item{radius}{}

\item{...}{Either \code{chemicals} or \code{chemicals, aerial_ground}. Not
quoted.}
}
\value{
A data frame with one row and the columns found in the
\code{calculate_exposure$exposure} data frame.
}
\description{
For a single date range, \code{row_out_df} returns a data frame with exposure
values calculated from \code{exp_out_val} as well as other relevant data.
This one row data frame is combined with data frames for other date ranges and
then returned as the \code{exposure} element from a \code{calculate_exposure}
list.
}
