% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-misc.R
\name{spdf_to_df}
\alias{spdf_to_df}
\title{Convert county SpatialPolygonsDataFrame to a tidy data frame.}
\usage{
spdf_to_df(spdf)
}
\arguments{
\item{spdf}{A SpatialPolygonsDataFrame object returned from
the \code{pull_spdf} function.}
}
\value{
A data frame with 24 columns if the \code{spdf} object is on the
section level and 23 columns if the \code{spdf} object is on the township
level.
}
\description{
\code{spdf_to_df} converts a SpatialPolygonsDataFrame object returned from
the \code{pull_spdf} function to a data frame.
}
\examples{
library(magrittr)
\donttest{
df <- readRDS(system.file("extdata", "fresno_spdf.rds", package = "purexposure")) \%>\%
   spdf_to_df()}
\donttest{
df <- pull_spdf("frenso") \%>\% spdf_to_df()
df2 <- pull_spdf("sonoma") \%>\% spdf_to_df()
# use df_plot() function to easily plot the output data frames:
df_plot(df)
df_plot(df2)
}
}
