% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-helpers.R
\name{tibble_to_vector}
\alias{tibble_to_vector}
\title{Return a character vector from a tibble with one column.}
\usage{
tibble_to_vector(tib)
}
\arguments{
\item{tib}{A tibble with only one column.}
}
\value{
A character vector.
}
\description{
\code{tibble_to_vector} takes a tibble with one column and returns the
values in that column as a character vector.
}
\details{
This is a helper function for \code{pull_raw_pur}, \code{pull_clean_pur},
and \code{pur_filt_df}.
}
